requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
fpdf2==2.7.6
openpyxl==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9